+++ tests/khrystynatk/hw5_khr_Page_object/base_page.py
@@ -1,5 +1,4 @@
 from selenium.webdriver.remote.webdriver import WebDriver
-from tests.khrystynatk.hw5_khr_Page_object.locators1 import LoginPageLoc
 
 
 class BasePage:
@@ -11,7 +10,5 @@ class BasePage:
     def open_page(self):
         self.driver.get(BasePage.HOST)
 
-    def input_creds(self, username, password):
-        self.driver.find_element(*LoginPageLoc.TXT_USERNAME).send_keys(username)
-        self.driver.find_element(*LoginPageLoc.TXT_PASSWORD).send_keys(password)
-        self.driver.find_element(*LoginPageLoc.BTN_LOGIN).click()
+    def get_url(self):
+        return self.driver.current_url
