+++ tests/khrystynatk/hw5_khr_Page_object/login_page.py
@@ -1,4 +1,5 @@
 from tests.khrystynatk.hw5_khr_Page_object.base_page import BasePage
+from tests.khrystynatk.hw5_khr_Page_object.locators1 import LoginPageLoc
 
 
 class LoginPage(BasePage):
@@ -6,5 +7,11 @@ class LoginPage(BasePage):
     def open_login_page(self):
         self.open_page()
 
-    def provide_creds(self):
+    def input_creds(self, username, password):
+        self.driver.find_element(*LoginPageLoc.TXT_USERNAME).send_keys(username)
+        self.driver.find_element(*LoginPageLoc.TXT_PASSWORD).send_keys(password)
+        self.driver.find_element(*LoginPageLoc.BTN_LOGIN).click()
+
+    def open_landing_page(self):
+        self.open_page()
         self.input_creds(self.USERNAME, self.PASSWORD)
